Remove watchdog for loading annotations. Use button instead
