+++ scripts/monitor/rano_monitor/handlers/__init__.py
@@ -1,7 +1,6 @@
 from .invalid_handler import InvalidHandler
 from .prompt_handler import PromptHandler
 from .report_handler import ReportHandler, ReportState
-from .reviewed_handler import ReviewedHandler
 from .tarball_reviewed_watchdog import TarballReviewedHandler
 
 __all__ = [
@@ -9,6 +8,5 @@ __all__ = [
     PromptHandler,
     ReportHandler,
     ReportState,
-    ReviewedHandler,
     TarballReviewedHandler,
 ]
